% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_caption_and_download}
\alias{add_caption_and_download}
\title{<img src="https://mint-vernetzt.de/static/e99e5a7a75c99c8651863585408242bb/mintvernetzt_og-img.png"',
<img src="www/MINTvernetztLogo_klein.png"',}
\usage{
add_caption_and_download(
  hc,
  filename = "plot.png",
  labelformat = "{point.y}",
  with_labels = TRUE,
  width = 450,
  height = 300
)
}
\description{
<img src="https://mint-vernetzt.de/static/e99e5a7a75c99c8651863585408242bb/mintvernetzt_og-img.png"',
<img src="www/MINTvernetztLogo_klein.png"',
}
