% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agriculture_data}
\alias{agriculture_data}
\title{Sample Agriculture Dataset}
\format{
A data frame with 50 rows and 2 variables:
\describe{
  \item{nitrogen_content}{Soil nitrogen content in mg/kg}
  \item{crop_yield}{Crop yield in kg per hectare}
}
}
\usage{
agriculture_data
}
\description{
A dataset containing crop yield and soil nitrogen content measurements.
}
\examples{
\dontrun{
data(agriculture_data)
linechart(nitrogen_content, crop_yield)
}
}
\keyword{datasets}
