% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{frequencies}
\alias{frequencies}
\title{Create frequency table}
\usage{
frequencies(var, data = NULL, percent = TRUE, sort = "desc")
}
\arguments{
\item{var}{Variable to tabulate (bare name)}

\item{data}{Optional data frame}

\item{percent}{Show percentages (default: TRUE)}

\item{sort}{Sort order: "desc", "asc", or "none"}
}
\value{
A data frame with frequencies
}
\description{
Create frequency table
}
\examples{
\dontrun{
mydata <- data.frame(category = sample(LETTERS[1:3], 100, replace = TRUE))
frequencies(category)
}
}
