% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_data}
\alias{heart_data}
\title{Sample Heart Disease Dataset}
\format{
A data frame with 200 rows and 5 variables:
\describe{
  \item{age}{Patient age in years}
  \item{sex}{Patient sex (Male/Female)}
  \item{chest_pain_type}{Type of chest pain (A/B/C/D)}
  \item{resting_bp}{Resting blood pressure in mm Hg}
  \item{cholesterol}{Serum cholesterol in mg/dl}
}
}
\usage{
heart_data
}
\description{
A dataset containing patient information and heart disease indicators.
}
\examples{
\dontrun{
data(heart_data)
piechart(chest_pain_type, sex)
}
}
\keyword{datasets}
