% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resolve_var}
\alias{resolve_var}
\title{Resolve variable name to column in data}
\usage{
resolve_var(var, data = NULL)
}
\arguments{
\item{var}{Bare variable name}

\item{data}{Data frame (defaults to finding any dataframe in environment)}
}
\value{
Vector of values
}
\description{
Resolve variable name to column in data
}
\keyword{internal}
